feat(github): have a seperate topic per event type
We will want to subscribe to some events but not all in our systems so
this will enable us to fan them out based on the event type.
We will need to figure out the attributes we want on each type of
event which will drive us to the right information at the right time.
for example we want to know action and actor for pull request

Signed-off-by: Jason Field <jason@avon-lea.co.uk>